requirements.txt
google-genai
tweepy
python-dotenv
requests
pytz